+++ src/model/invite.rs
@@ -279,7 +279,7 @@ impl InviteGuild {
     /// ```
     #[cfg(all(feature = "utils", not(feature = "cache")))]
     #[inline]
-    pub fn shard_id(&self, shard_count: u64) -> u64 {
+    pub fn shard_id(&self, shard_count: u32) -> u32 {
         self.id.shard_id(shard_count)
     }
 }
